+++ spec/features/dependency_observe_spec.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-RSpec.describe 'Dependency observers (#observe, #unobserve, #clear_listeners)' do
-end